Fix availability triggers on client error
Details
In the context of intra-ECU communication, if a client loses the
connection to a service provider, then it should re-request service
availability triggers from the routing host for the provided services.
This would allow the client to handle the error and re-establish the
connection to the service provider, however, it does not appear that
this mechanism is working correctly.
These changes attempt to fix the issue by searching through
local_services_ for the services offered by the provider, instead of
searching through the services to which the client is subscribed.