Notify stop func when receives EOF during stop process
Summary
Synchronization issue found during the stop process when one side of the connection closes
earlier than the other, allowing the port to be reused while there is still an active connection.
Details
This issue occurs in a very specific scenario:

The application starts the stop process, closing connections with routing manager, and routing manager
also starts closing connections with the application.
routing manager still has messages in the queue on the routing manager client -> application server connection,
waiting in 10ms cycles for the queue to become empty.
The application closes the application server -> routing manager client connection and sends an EOF to routing
manager, making port XXXXX available for new connections.
A new application connects to port XXXXX because it is free, but fails to connect because routing manager
still has an active connection to port XXXXX.
After the error, it enters a loop and cannot recover.

Issue 1: A port is assigned to a new application but there is still a connection to that port for
the previous application
A condition variable was implemented to notify the stop process to exit the 10ms wait if an EOF is received
in receive_cbk, since the queue is cleared in this case.

local_tcp_client_endpoint_impl.hpp: Added bool is_stopping_ and condition variable queue_cv_.
local_tcp_client_endpoint_impl.cpp: Updated receive_cbk to notify stop() if the stop process is in progress
and an EOF is received.
Updated stop() to wait for either 10ms or until the queue is empty, instead of just waiting 10ms.

It was possible to reproduce this issue and with this change, it was possible to immediately close
the socket to prevent the issue from occurring.

Issue 2: The application does not recover the connection after the register timeout
client_endpoint_impl::cancel_and_connect_cbk was updated to only report the connection as successfully
terminated when cancel_and_connect_cbk is called without error. Previously, it was reporting the connection
as successfully terminated even when there was a connection error, preventing the endpoint from being
restarted.
It was possible to reproduce this issue and with this change, the issue was resolved.
After the register timeout, the client can now recover the connection.