Ensure "shutdown_and_close" is called on "stop"
Summary
Ensure "shutdown_and_close" is called on "stop"
Details
A call to "stop" sets the "is_stopped_" flag and cancels the running operations on the
sockets of the endpoint. In case the operations are scheduled but not yet running,
their callbacks are called with "operation_aborted" and call "shutdown_and_close",
which handles the sockets and, in case of a restart, reinitializes and starts the endpoint.
In case the operation is already running (executes its callback) at the point in time
"cancel" is called, we currently do detect the endpoint is stopped by checking the flag,
but do not call "shutdown_and_close". As "shutdown_and_close" is not triggered, the
endpoint will stay stopped and can also not be restarted by another call to "restart",
because no operation was scheduled that could be cancelled. To avoid this, this change
also calls "shutdown_and_close" on the regular path of the callbacks.