set TCP_USER_TIMEOUT on local connections
Details
When an endpoint is no longer reachable (after a suspend, for example),
how quickly the TCP connection takes to timeout depends on its' state:

with no unacknowledged data, it depends on TCP keepalive, and depends
on the underlying kernel configuration (interval, number-of-probes, ...)
with unacknowledged data, it depends on the configuration for TCP
transmission, the number of retries, and exponential timeouts

It is known that devices sometimes  have very aggressive
TCP keepalive parameters (3 probes, 1 sec).
The TCP retransmission parameters however are usually default, and it can
take many, many, many minutes until a TCP connections due to
retransmission exhastion
This can lead to  Android <-> Node0
connections to timeout quickly, after ~3s, when Android suspends. However,
connections which happen to have a lot of data flowing through (and thus
unacknowledged data), would often never timeout. This is problematic, as:
a) once Android resumes, those those "stale" connections need a TCP
retransmission to recover, which can take up to 2 minutes
b) in the meantime, VSIP was accumulating gigantic buffers, risking a OOM
In some cases, it even seems deterministic - connection between VSIP and
a app would never timeout after Android suspends, due to the sheer volume
of data that goes through
To cope with this gigantic mess, use TCP_USER_TIMEOUT, which causes a
connection to timeout when data remains unacknowledged for more than
than a certain amount of time
There are many, many advantages to TCP_USER_TIMEOUT:

it is extremely granular, can set timeouts in the order of
milliseconds
(which might be interesting for vsomeip, as all communication is local)
it overrides both TCP keepalive and TCP transmission mechanisms, such
that it becomes easier to reason about when a connection will timeout
it ties quite nicely with TCP keepalive, which ensures that there is
periodically some data to acknowledge
(without TCP keepalive, a connection would otherwise remain stale)

While at it, disable keep alive for same-SoC/VM connections