Test for registration process
Details
The Out of Memory condition that triggers the yellowscreen happens when clients
repeatedly attempt to reconnect.
The root cause occurs when the TCP ACK for the registration message is delayed,
leading clients to continuously retry the connection.
In this test, a service application (registration_check_service.cpp) attempts to register
with the routing manager.
During this process, a Python script (someip_sequence_checker.py) is activated to
intercept outgoing TCP packets using NetfilterQueue and Scapy, monitoring for a
specific sequence of SOMEIP protocol messages during the registration phase.