Handle not_socket error on send_cbk
Summary
Resolve the issue of the socket not being able to be opened after Resume due to the
"Socket operation on not-socket" error
Details
For some reason, when entering suspend mode, instead of the file descriptor being deleted by the kernel
as happens in normal cases, the file is not being deleted and the socket is only being marked as invalid.
In the case of the "bad file descriptor" error, the socket is shut down and reconnected to establish the new socket.
To resolve the "Socket operation on not-socket" issue, an error handler is implemented in send_cbk with the same
mechanism as "Bad file descriptor" since the error type is similar - instead of the file being deleted when entering
suspend, it couldn't delete the file but marked the socket as invalid. The end goal is the same.