Set was_not_connected to true if not connected
Summary
Set was_not_connected to true if not established or connected
Details
Previously, if a tcp connection was not opened when trying to send
a packet, the send_cbk function would just set the is_sending_ flag
to false, without any mechanism to retransmit the message.
By setting the was_not_connected to true in this situation,
when the connection is established the message will be sent again
in the connect_cbk function.