Disconnect on error
Summary
Close all connections on "client error".
Details
These changes related to e716737 but here it's the server that loses
connection to the client.
Since the server application may not be consuming services from the
client, the process of re-requesting service availability triggers from
the routing host does not work.
Instead, the server will forcibly terminate all connections to the
client, forcing them to re-establish the connections to the server.
An alternative was considered  where the server would instead
try to reconnect to the client when it tried to send something, but it was
a bigger change that introduced several issues during testing.