vsomeip examples on Windows not running
Summary
Updated utility::is_routing_manager and documentation for Windows
Details
When a process is interrupted midway (e.g. kill -9), the Host file is not deleted and when attempting
to launch a new host, since the file already exists, it's not possible for the Host to be created, never being
able to have a new Host again. A handler was created for Windows to, in case it cannot create the file, verify
if it's still valid and if not, delete the file and create a new one. In the case of Linux, it can already verify if the
file is locked or not, and if it's not, it assigns the new host to the file.
Additionally, documentation was updated for Windows:

README.md
examples/hello_world/readme
examples/readme.txt changed to readme.md and updated