rmc: drop use of netlink
Details
With TCP communication, netlink is also used by every vsomeip
application in order to determine when the "local" interface is ready
That unfortunately means that an ECU with a few hundreds of vsomeip
applications has a few hundreds of netlink sockets, all racing at
startup to collect network state. Not ideal!
Instead of using netlink (which is Linux-specific), use IP_FREEBIND
(also Linux-specific) in order to bind the local TCP ports regardless of
network state. It is much simpler, and bypasses the problem of waiting
for the local interface
There are also plenty of small other advantages:

routing client logic becomes similar in TCP/UDS
ton less code