Always send initial event on subscription
Summary
Always send initial event on subscription.
Details
This removes a check from the subscription logic for internal
communication which prevented notifications from being sent if a client
re-subscribed to an active subscription.
The issue was that in the case where the subscription was not properly
canceled, such as during a suspend cycle, the new subscription would not
trigger the initial events that the client was expecting.
Since the client is the one interested in maintaining the subscription,
we should consider it the authority regarding the subscription status
and treat its requests accordingly.