Update security offer rule for messages
Update security offer rule for messages
Details
With the introduction of 0622382, an issue in the policies was detected affecting internal security in the
following scenario:
1- Client A connects to the daemon and receives the policies loaded from
vsomeip_security.json and <uid>_<gid>/vsomeip_security.json.
2- Client B connects to the daemon and wants to communicate with Client A.
3- The daemon sends an UPDATE_SECURITY_CREDENTIALS_ID message to Client A with Client B credentials.
4- If Client A does not yet have the credentials for Client B, it adds the credentials to any_client_policies_.
Until now this worked well because it only verified if the client was allowed and we only added the UID/GID
to the allowed clients list. Now since it verifies that the offer is allowed, it doesn't work because the policy load
wasn't done.