Fix increase in exported symbols
This commit 39f6770 mistakenly broke the ability to hide export symbols.
Here the functionality is fixed.
Details
The number of exported symbols increased exponentially due to the recent
CMakelists update. The export symbols flag were being removed, and that was resulting
in this increase.
Here the unneeded optimization flag was
removed