Change order of local endpoint clean-up
Summary
Change the order how a connection is cleaned-up.
Details
When a server -> client connection is broken the endpoint needs to be cleaned up,
before the corresponding client -> server connection is closed.
Otherwise it can happen that the later is restored and the former tried to be used to
acknowledge a subscription.