clear multicast only for last subscriber
Summary
Check if last subscriber was removed before clearing multicast endpoint
Details
When two clients are subscribed to a multicast eventgroup,
the function clear_multicast_endpoints was
called when one of the clients unsubscribed to the eventgroup,
even if there were other clients still subscribed to the multicast
eventgroup.
This lead to a situation where the endpoint leaved the multicast
group and the clients that were still subscribed did not get
notifications sent by the service until the multicast group was rejoined.
To fix this situation, now the unsubscribe function checks if the
client is the last subscriber to the eventgroup.