flaky test someip_application_init_test.multithread_init
Summary
Fix test someip_application_init_test.multithread_init
Details
It waits for all applications to be initialised before exiting.
This is necessary because one application is the host and
must remain present during the initilialization of the
proxy applications. Else a proxy application will try
to connect to the host and it will fail.