Solve availability test deadlock
Fix availability handler client deadlock.

Sporadically, the availability tests failed by timeout since the client
application didn't terminate properly and continued on loop trying to
connect to host. The client shutdown thread was blocked waiting for the
dispatch thread which was trying to acquire the client inter-process
mutex in the availability callback, which was held by the client in the
stop method.

Why sporadic? It depends on (non-deterministic) locking order of client
mutex between the dispatch thread and the stop method.