rms: drop use of netlink
With TCP communication, netlink is used by a host vsomeip application to
determine when the "local" interface is ready.

Instead of using netlink (which is Linux-specific), use IP_FREEBIND
(also Linux-specific) in order to bind the local TCP ports regardless of
network state. It is much simpler, and bypasses the problem of waiting
for the local interface.

Same is done in rmc, see e8b6519eb8f0035d704a1a99f5fc4532a12b8117