control access to targets_
Ensures that all accesses to targets_ are synchronised by mutex_.

There is a segfault when usei::send_queued_unlocked performs the following
operation:
const auto its_entry = _it->second.queue_.front();

The back trace is
\#1  __gnu_cxx::__atomic_add_dispatch (__val=1, __mem=0x402e2cbc6794f03)
\#2  std::_Sp_counted_base<(__gnu_cxx::_Lock_policy)2>::_M_add_ref_copy (this=0x402e2cbc6794efb)
\#3  std::__shared_count<(__gnu_cxx::_Lock_policy)2>::__shared_count (__r=..., this=<synthetic pointer>)
\#4  std::__shared_ptr<std::vector<unsigned char, std::allocator<unsigned char> >, (__gnu_cxx::_Lock_policy)2>::__shared_ptr (this=<synthetic pointer>)
\#5  std::shared_ptr<std::vector<unsigned char, std::allocator<unsigned char> > >::shared_ptr (this=<synthetic pointer>)
\#6  std::pair<std::shared_ptr<std::vector<unsigned char, std::allocator<unsigned char> > >, unsigned int>::pair (this=<synthetic pointer>)
\#7  vsomeip_v3::udp_server_endpoint_impl::send_queued_unlocked (this=this@entry=0x7f84002180, _it=...)
\#8  vsomeip_v3::udp_server_endpoint_impl::send_queued (this=0x7f84002180, _it=...)
\#9  vsomeip_v3::server_endpoint_impl<boost::asio::ip::udp>::send_cbk (this=0x7f84002180,

This means that it, which is of type target_data_iterator_type, pointed to
corrupted data. It had been retrieved from server_endpoint_impl::send_cbk,
which obtained it as follows
    std::lock_guard<std::mutex> its_lock(mutex_);
    auto it = targets_.find(_key);
    if (it == targets_.end())
        return;

If all accesses to targets_ are synchronised using mutex_, it should not be
possible to corrupt the it pointer. This is what is done in this PR.