Check availability on send
Check if a service is available before sending a request to it.

This would prevent us from trying to process a request which might not
reach its destination, and allows us to immediately report an error to
the application.

Process offer service before on_availability callback being invoked,
with previous order, clients could send a request while the service is
still not added to the control map.

Npdu test is updated to enforce the clients to request services before
sending any request type message, without the request, the host will not
send the ADD_SERVICE routing info command and consequently on the client
side the service will constantly be assessed as unavailable.