fix regression introduced by usei refactoring
Check that connection wasn't stopped before to repeat the receive.

After refactoring udp_server_endpoint_impl, there is a lifecycle_index
which is used to check whether the result of an asynchronous operation
must be ignored. The access to this variable cannot be synchronized
because  the call to on_message_received_unlocked must be lock-free.

But after calling on_message_received_unlocked, we must check again
the value of lifecycle_index before to repeat the asynchronous
operation. And this second check was missing.

Without this check, the asynchronous operation can be repeated, even if
a stop was requested in between.