udp-server: fix automatic rejoin operations at startup
The function that automatically restored joins at startup was defective.

The function was faulty because it was performing invalid checks. The
need behind this function is unknown. It could be a performance
optimization. There were two options: delete it because it seems
redundant, or fix it. The second option has been chosen.

This function checked whether the group was joined successfully before.
But after a stop, this information is no longer valid. This check has
been removed.

It also checked whether the socket has been created previously. However,
the socket may have been created by set_multicast_option, in which case
we still need to join the groups. This check has been removed as well.
We can remove both checks because redundant join operations are
ignored by set_multicast_option.