Mitigate to add already existing connection
Problem scenario:

Client successfully connects for Routing HOST;
Before the client starts the Registration process it receives EOF
(unknown reasons);
Client retries to reconnect with HOST, however host denies because its
connection with this client already exists;
Client gets stuck in this loop.

The main issue is that, because the TCP keep-alive is disabled, for
clients in the same VM/SOC, the HOST never discovers that it lost the
connection.

With keep-alive enabled, this process is recoverable, HOWEVER,
it takes more time, equal to what are the tcp_keekaplive settings.

Solution:

The proposed is to simply replace the previous connection, as it is a
faster recovery. Also adds test to replicate the issue. Adds to the
socket_manager the ability to allow the disconnection of  only one side
of the endpoint.