fix test suspend_resume_test_initial
Fix race condition in test suspend_resume_test_initial when a notification is received after the unavailability.
the test failed because a notification was received when the service was unavailable.
2025-08-11 14:44:47.829423 suspend_resume_test_service [info] rmi::set_routing_state
      Set routing to suspend mode, diagnosis mode is inactive.
2025-08-11 14:44:47.829768 suspend_resume_test_client [debug] on_message: Received event.
2025-08-11 14:44:47.829864 suspend_resume_test_client [debug] on_availability: Test service is NOT available.
2025-08-11 14:44:47.829890 suspend_resume_test_client [debug] on_message: Received event.
2025-08-11 14:44:47.829894 suspend_resume_test_client [debug] [TEST-cli] HasReceived Changed, triggering cv

This last trace means that the variable has_received_ has been set to true.
After resume, the test expected to receive a notification, which it did. But this was ignored because the
variable has_received_ was not set to false.
2025-08-11 14:44:52.829535 suspend_resume_test_service [info] rmi::set_routing_state
      Set routing to resume mode, diagnosis mode was inactive.
2025-08-11 14:44:52.840545 suspend_resume_test_client [debug] [TEST-cli] On availability will trigger cv
2025-08-11 14:44:52.840596 suspend_resume_test_client [debug] [TEST-cli] Service Available after susp/resume: r=0
2025-08-11 14:44:54.836073 suspend_resume_test_client [debug] on_message: Received event.

There is a defect in the test that requested two notifications by generating unsubscribe/subscribe commands
without waiting for confirmation of these operations. The result of these two operations interfered with the
expected sequence.