Force remote subscriptions to be removed on host error
Force remote subscriptions to be removed on host error.
In some ECU modes, an emergency shutdown is executed instead of a graceful shutdown,
this leads to vsomeipd not suspending, consequently, the vsomeip suspend
command is not propagated to clients.
For some clients, as the suspend command is not received and the unit
doesn't reboot, it will not clean up the remote subscriptions map that it manages,
the container is used to detect if a client is a new subscriber or not to an event (initial
events are only sent if it's a new subscriber).
This PR forces the clean up for scenarios where a graceful shutdown is not performed by
cleaning up the container when the client detects and handles a connection lost towards host.