ltcei: fix connect logic
Due to e8b6519eb8f0035d704a1a99f5fc4532a12b8117 (netlink removal), libvsomeip receives a EADDRNOTAVAIL
on bind, in ltcei::connect, before the network interface is ready
Which by itself is expected, and not a problem - but due to broken error
handling, libvsomeip continues with connect logic, does ASSIGN_CLIENT
The sending of ASSIGN_CLIENT of course fails (connection is not open..),
and causes a cascade of errors, which unfortunately only resolve on the
ASSIGN_CLIENT_ACK timeout, which takes 3s and heals the situation.
There are failures exactly because of these extra 3s.
While at it, promote ASSIGN_CLIENT/REGISTER_APPLICATION timeouts to
errors, as these should never happen