multicast, add repeat delay
Avoid a busy loop if a failure occurs during a multicast leave operation.
The leave failed because MACsec reconfigured the network interface
during the normal execution. The endpoint_manager_impl doesn't
repeat the leave but it repeats the join in a busy loop.
By performing the leave operation even if it was already done,
it will avoid the error: address already in use, when the join is
done again.
The error: address already in use, cannot be ignored because
it happens during the restart case and it must be managed in
this case.
It isn't enough to fully fix the busy loop, because the join
will return an error until the network interface is up again. So
a delay has been added,
It remains one issue: the leave isn't repeated by
endpoint_manager_impl. Only the join is repeated.
The correct handling of this problem would require to record that
the join state as unknown. Then we would have to manage this
new state and repeat of the leave operation. It is too risky for
something that shall not happen.