Update local_endpoint ClientID on add_guest
Update local_endpoint ClientID on add_guest
This was the case:
There was an endpoint established between two applications A and B.
An STR occurred.
The endpoint still exists, but is disconnected. It tries to reconnect, and succeeds.
During this reconnection, application B re-registers to routing host and changes ClientID.
Afterwards app A receives ON_AVAILABLE from app B, with a new ClientID.
App A tries to connect to this new ClientID, which has the same address, and it fails with
Cannot assign requested address.
This fix proposes to update the local_endpoint map, to synchronize the ClientID with
what we are updating in the guests map (when 'add_guest' is called).