add SO_REUSEPORT option + logging
SO_REUSEPORT option to mitigate the issue with Address already used.
After resume, the TCP server that was closed during suspend received
an unexpected error: Address already in use. The reason is not clear,
and the DLT log shows that the stop was called. The workaround is to
use the option SO_REUSEPORT.
The same option is added to the UDP unicast server. The UDP multicast
server already has error handling and does not require it.
This PR also aims to improve logging, particularly by tracking init,
start and stop operations, and by monitoring active connections.