Fix external_local_payload_test_client_external
Temporary fix for external_local_payload_test_client_external
This test keeps failing on memcheck, likely due to a timing issue introduced by memcheck
slowing down the execution
The premise of the test is to have external communication while SD is disabled,
by creating the endpoints directly from the config file of the service.
What happens is that when the client application starts before the server,
the client will immediately try to establish a connection even before the service is offered.
Once the service is finally offered, it triggers the execution of two availability handlers,
despite only one being registered. This then leads to a BLOCKING CALL and the test gets stuck,
the exact reason is still unknown to me.
If we force the server to always start before the client, the test never presents this failure
While there is an underlying issue found in this particular scenario, which should be fixed
and will continue to be investigated. I propose merging this change so that the
test is reliable and while a deeper analysis is conducted