credentials: fix missing MSG_NOSIGNAL
SIGPIPE is one of the quirkier POSIX details. If a process does a
write/sendmsg to a pipe that shutdown/closed, it receives a
SIGPIPE
The codebase generally does not have to worry, because boost asio deals
with it. However, credentials does a sendmsg manually, and does not
use MSG_NOSIGNAL, which is what prevents generation of a SIGPIPE