Adjust registration state before offerings
Adjust registration state before offerings
The sequence that caused this race condition to cause failures were:

A service started and listed three services each with three fields to be offered
The service had initially trouble connecting to the router
A client of the service requested this service and subscriptions for the fields
The service managed to connect to the router
The service send the ack
The service iterated over all the pending offers (it did not change the internal state yet - still in REGISTERING)
The service started to offer three fields for one of the services
The service received the subscription of the client while still iterating over pending offers to send
when trying to send the initial field, the sending would fail, because the internal state was not yet "REGISTERED"
The sub_ack was send successful, because there is no such check for this command

To resolve this issue it was chosen to send the pending offers as late as possible, while still holding the register
mutex, to ensure that the sending of pending offers would not interleave with a subsequent de-registration.
Although it remains questionable that the pending offers are send out before the vsomeip-application
callback for being registered is invoked.