add_guest only updates rmc
Follow up of
The add_guest updating the local_endpoints_ interferes with the (de)registration process.
This change reverts the add_guest for the old functionality, only for the routing host.
This topic needs to be revisited.
During STR, if a client sends a deregistration, and the host only processes it on resume,
the client will always fail the first re registration attempt.