Add basic watchdog for tracking stop offers
Basic watchdog to track if for all externally offered services a stop offer command is sent
to the network.
The stop offer watchdog is implemented in the SD library, when externally offered services
are "collected" on suspend and sent to SD module to manage the stop offer commands, the
watchdog, if enabled, is triggered for the configured timeout.
Upon trigger, it iterates through a copy of the collected services to guarantee all have
corresponding stop offer commands sent to the network. The detection of stop offer commands
is done via the sent unicast/multicast traffic callback on SD endpoint.