Update cache on change
In order to reduce network load, vsomeip caches previously received
initial events which can be reused when a new client subscribes.
This change fixes an issue where the interface version was not
being correctly cached/updated.
Note
This is a follow-up to this discussion that seeks to avoid the use of
atomics.