remove sessionID from clients_ map
remove sessionID from clients_ map
Repeating references to the same endpoint were being saved
in clients_ map due to it considering the sessionid. Since the
sessionid is incrementing with each new message sent, it would
lead to a huge map with repeating entries.

This and 545ca62bc5ceb6367adcb18e6dcd5fea203d7921 made the memory_test fail
because it was causing a big and sudden increase in memory
usage.