Fix subscribe notify test one tests
Fix subscribe notify test one tests by adding adicional
checks for initial events and clear received events when
setting a new payload
Before, the subscribe notify tests were failing due to the "issue"
with initial events being sent twice very sporadically, This however
is not a real issue and as such should not lead to the test failing.
As such, the test should pass if more than one initial events is sent
for a given eventgroup. For that purpose, new checks for the initial
events were implemented to accommodate for the situation of
double initial events being sent for either or both of the eventgroups.
Also, clear the received events map when receiving the response from
the server to the set payload message, since if the double initial
events occur the received events would have messages with
the previous payload still stored.