Check ACL and process messages if the address is specified
Only check the ACL and process messages if the address is specified
Update the ´routing_manager_impl::on_message´ to ensure that the ACL check is only performed if the remote
address is not "unspecified" (e.g., 0.0.0.0). If the address is unknown, the message is ignored and a warning is logged.
Additionally, it fixes the ACL rejection log to correctly display the service and instance, instead of repeating
the instance value.
It also removes commented and redundant code related to ACL checks for local testing.