Add boardnet initial event test
Test the remote subscription clean up on client to host re-connection

When an emergency shutdown is executed instead of a graceful shutdown,
this leads the host not receiving a SUSPEND message, consequently, the host suspend
command is not propagated to both local clients.
For some clients, as the suspend command is not received and the unit
doesn't reboot, it will not clean up the remote subscriptions map that it manages,
the container is used to detect if a client is a new subscriber or not to an event (initial
events are only sent if it's a new subscriber).
This test is to ensure that fix introduced in eb10710f56524d64de6c09f53b7653e5ab794601
works and the remote subscriptions are removed on client to host re-connection.