manage suspend event
Stop servers during suspend to avoid dangling connections.
Recently, the Ethernet interface has been deactivated during
suspend, and at the same time, unexplained errors have
appeared: device not found, address already used and cannot
assign requested address.
For instance, when leaving the multicast group the following
error was reported: Cannot assign requested address.
This is probably a consequence of the network interface being
deactivated during suspend. Linux is no longer able to send
periodic IGMP messages to join the multicast group, and it
leaves the group without informing the application.
It is probablematic because it means that the vsomeip
wasn't anymore part of the multicast group without being
informed.
The fix is to stop servers offering services to other ECUs during
suspend, eliminating the risk of desynchronisation between
applications and Linux due to the network interface being
deactived.
Another correction was made to force the execution of the
suspended state, when there is a transition from running to
resumed.