application start-stop-start broken with local TCP
Fix routing_manager_client::start()
If a new application is started between the stop and the second start of the first application, the start of the
first application will fail. This happens because, since 'create_local_server' is not called in
routing_manager_client::start(), the server endpoint will be assigned, in 'create_local', to a memory port
that is already being used by application 2.
Scenario:

app1->start() - port 14321
app1->stop() - endpoints destroyed and ports released
app2->start() - port 14321
app1->start() - port 14321 because create_local_server is not called -> registration fails because the address
is already assigned

To solve this problem, the creation of the receiver_ endpoint was added in routing_manager_client::start()
to ensure that a free port is always assigned.
Additionally, during this process the sec_client_.port became desynchronized from the server port because
set_sec_client_port is only set in routing_manager_client::init(). It was found that this set_sec_client_port is
duplicated since it is also set in endpoint_manager_base::create_local_server, so it can be removed.
By moving create_local_server from routing_manager_client::init() to routing_manager_client::start(), it
is ensured that sec_client_port is always set with the correct port.