Wait for unacknowledged data before closing
Wait for unacknowledged data on all TCP endpoints

Before calling close() add a wait for all data to be sent and acked.

Setting the linger option, on tcp sockets, will not change the TIME_WAIT value.
The linger will block the close call, for a maximum of the defined time.
This means, if we set linger to 5, the close() call can block up to 5 seconds.
Afterwards, the socket can still be in a TIME_WAIT state for 60~120 seconds.
The only way to fully remove the TIME_WAIT state is to set the linger to 0, which forces a RST to be sent.
The issue with this approach is that when calling the close() the RST will be sent and the buffered data,
that is sill to be sent, will be dropped.
In order to avoid this, before calling the close(), vsomeip-lib needs to wait until all data is sent.