extend multigroup test to detect issue with UDP dispatching
Modify the multicast group test to detect issues related to UDP packets dispatching.
SO_REUSEPORT was added to workaround a TCP issue related to the restart.
During the review process, it was added too to UDP but it introduced a
regression: UDP packets were no longer being received by the correct socket.
Indeed, there are no written rules defining how UDP packets should be
distributed by Linux when multiple sockets are bound to the same endpoint.
VSOMEIP can create up to three sockets on the same endpoint. It was
thought that when the SO_REUSEPORT option was enabled, Linux routed
all traffic to the one bound to INADDR_ANY. That is not correct. Therefore,
the fix that redirected messages received by the socket bound to
INADDR_ANY has been removed from this PR, to avoid any regression. It
remains a check that prints a warning if this socket receives a packet
addressed to another socket.
Therefore, limited to extending the multicast group test in order to
detect the problem of unexpected dispatching of UDP packets.