Update error handling for local endpoints
Update the error handling for local communication to synchronize IO operations errors and state.
In some scenarios, a race condition between
local client endpoints read and write callbacks error handling could force an application to never
fully close the socket and be stuck on a re-connection loop.
This PR updates the error handling of both IO operations for local tcp and uds communication.
The callbacks for a single connection are invoked sequentially and only the first one is processed,
moreover, the re-connect decision is forwarded for the routing manager instead of forcing low-level
retries.