restrict routing state to routing_manager_impl
Preparatory step in refactoring routing manager states.
The routing_state_e is not really a state, but rather an
event that informs the routing manager of a lifecycle-related
event. The end goal is to correct this.
As a preliminary step, it will be removed from
routing_manager_client that currently ignores it.