Change return type of get_sec_client
Change return type of get_sec_client.
The former signature was not thread-safe.
Granting any access to internal
memory, when the content of this memory could change
(set_sec_client_port), is not thread safe.
At the moment the memory is guarded by the higher order protocol
(a port is only set during the init phase) for which
reason this is not a practical problem at the moment.
But changing the port only during the "init" is at least a
theoretically problem, as the port is temporarily freed
on a reconnect and could be "stolen".