Clear accept handlers on io_stop
In most cases this is not necessary, as the receiver_->stop() will lead to a deletion.
But if a reconnect is attempted, during stop the receiver_ can be restarted.
To ensure that the socket_manager breaks the circle (as io_context would usually do),
also the acceptor handlers should be cleared on io_stop.