suspend_resume_test_initial - Detect StopOffers
Detect the timing of stop offers during STR
suspend_resume_test_client.cpp updated to detect when unavailability is detected and measure the time until
receiving the next availability to verify if the StopOffer was sent during suspend or resume. We cannot read the
messages directly in the test, but considering that the network is down for at least 2 seconds, if the time between
unavailable and available is less than 2 seconds, it can be considered that the StopOffer was sent during Resume
(all tests performed showed that whenever the StopOffer is sent during resume, the time between unavailable and
available is almost instantaneous, with the highest observed value being 300ms).
This change can be important to detect problems introduced in the suspend process in the future.
Note: With the current state of the suspend process, considering that sending StopOffers is asynchronous and we
don't have full control over the Suspend process, the test sporadically fails because in some STRs the server endpoints
end up being closed before the StopOffer is sent.